organize functions better between introspection contexts and util files for KSP and reflection
